# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tech Mentor ERP & POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: integer } }
        - { name: brandId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/import:
    post:
      operationId: importProducts
      tags: [products]
      summary: Import products from Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # BOOK CATALOG
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, salePrice, costPrice, stock, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        barcode: { type: ["string", "null"] }
        isbn: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        subCategoryId: { type: ["integer", "null"] }
        subCategoryName: { type: ["string", "null"] }
        branchId: { type: ["integer", "null"] }
        branchName: { type: ["string", "null"] }
        callNumber: { type: ["string", "null"] }
        brandId: { type: ["integer", "null"] }
        brandName: { type: ["string", "null"] }
        seriesId: { type: ["integer", "null"] }
        seriesName: { type: ["string", "null"] }
        classId: { type: ["integer", "null"] }
        className: { type: ["string", "null"] }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        author: { type: ["string", "null"] }
        edition: { type: ["string", "null"] }
        costPrice: { type: number }
        salePrice: { type: number }
        discountPrice: { type: ["number", "null"] }
        stock: { type: integer }
        lowStockLimit: { type: integer }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        status: { type: string }
        unit: { type: string }
        batchNumber: { type: ["string", "null"] }
        mfgDate: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, sku, salePrice, costPrice]
      properties:
        name: { type: string }
        sku: { type: string }
        barcode: { type: string }
        isbn: { type: string }
        categoryId: { type: integer }
        subCategoryId: { type: integer }
        branchId: { type: integer }
        callNumber: { type: string }
        brandId: { type: integer }
        seriesId: { type: integer }
        classId: { type: integer }
        subjectId: { type: integer }
        author: { type: string }
        edition: { type: string }
        costPrice: { type: number }
        salePrice: { type: number }
        discountPrice: { type: number }
        stock: { type: integer }
        lowStockLimit: { type: integer }
        description: { type: string }
        imageUrl: { type: string }
        status: { type: string }
        unit: { type: string }
        batchNumber: { type: string }
        mfgDate: { type: string }
        expiryDate: { type: string }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        barcode: { type: string }
        isbn: { type: string }
        categoryId: { type: integer }
        subCategoryId: { type: integer, nullable: true }
        branchId: { type: integer, nullable: true }
        callNumber: { type: string, nullable: true }
        brandId: { type: integer }
        seriesId: { type: integer }
        classId: { type: integer }
        subjectId: { type: integer }
        author: { type: string }
        edition: { type: string }
        costPrice: { type: number }
        salePrice: { type: number }
        discountPrice: { type: number }
        stock: { type: integer }
        lowStockLimit: { type: integer }
        description: { type: string }
        imageUrl: { type: string }
        status: { type: string }
        unit: { type: string }
        batchNumber: { type: string }
        mfgDate: { type: string }
        expiryDate: { type: string }
    ImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
        errors:
          type: array
          items:
            type: string

    # CATEGORY
...